Move tests from tests/__init__ so unittest2's discover can find them
--HG--
rename : rst2rst/tests/__init__.py => rst2rst/tests/test_fixtures.py
